refactor node_editor; extract form handling -also move csrf token logic to api
